Add CircleCI configuration
At the moment it doesn't seem particularly easy to work with it with Clojure / JDKs so it won't replace GH Actions.

Still it's a worthwhile addition since it offers greater build concurrency and available CPUs.

So one can monitor https://app.circleci.com/pipelines/github/reducecombine for a greater feedback cycle (I have that repo as an additional `pushurl`).

Closes https://github.com/clojure-emacs/enrich-classpath/issues/4